fix Rswiftの拡張を別ファイルで行うようにした。
refs #8
